+++ eth/handler_test.go
@@ -605,7 +605,7 @@ func testBroadcastBlock(t *testing.T, totalPeers, broadcastExpected int) {
 			}
 		}(peer)
 	}
-	timeout := time.After(300 * time.Millisecond)
+	timeout := time.After(600 * time.Millisecond)
 	var receivedCount int
 outer:
 	for {
